split tools, add call for internal tools, prepare toolkits
